Refactor DocumentMergeConflict into own file
